feat: Reorder navigation buttons and redirect to chat view
